编解码工作完成

+++ rpc/src/main/java/org/minidubbo/rpc/codec/Serialization.java
@@ -1,7 +1,7 @@
 package org.minidubbo.rpc.codec;
 
 public interface Serialization {
-    int FASTJSON = 0;
+    byte FASTJSON = 0;
 
     byte getSerializationType();
